% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addError.R
\name{addError}
\alias{addError}
\title{Adds error radius to point coordinate data}
\usage{
addError(x)
}
\arguments{
\item{x}{An object of sf, sfc or sfg format, containing a geometry
and an Extent_km column}
}
\value{
returns \code{x} with point radius method applied to geometry column,
turning longitude/lattitude point data into a polygon of \code{Extent_km}
raidus centred on the point data given in \code{geometry}
}
\description{
\code{addError} adds an error radius using the point-radius method outlined
in this here \link{http://www.herpnet.org/herpnet/documents/wieczorek.pdf}
and converts to data to a long/lat CRS.
}
\examples{
data.frame <- addError(data.frame)
}
