% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{bound}
\alias{bound}
\title{Used to get the min/max of x/y dimension of polygon}
\usage{
bound(x, dim, type)
}
\arguments{
\item{x}{sf object}

\item{dim}{dimension, 'X' or 'Y'}

\item{type}{type of computation, 'min' or 'max'}
}
\description{
Used to get the min/max of x/y dimension of polygon
}
