% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroidCentroidDistance.R
\name{calculateCentroidDistance}
\alias{calculateCentroidDistance}
\title{Calculates the distance between the centroids of two convex hulls}
\usage{
calculateCentroidDistance(x, y)
}
\arguments{
\item{x}{an object of class sf, sfc or sfg containing data from Natural
History Museum archives. Contains a \code{binomial} column and geometry.}

\item{y}{n object of class sf, sfc or sfg containing data from IUCN records.
Contains a \code{binomial} column and geometry.}
}
\value{
returns \code{x} now with additional \code{distance} columns, which
contain the computed distance between the centroid given in \code{x} and the
centroid given in \code{y}
}
\description{
\code{centroidCentroidDistance} calculates the distance from the centroid of one
convex hull to the edge of up to two other hulls for a given species. This
was developed to produce data for further analysis of species range changes
from historical records to current estimates.
}
\examples{
x <- centroidCentroidDistance(x, y)
}
