% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateOverlaps.R
\name{calculateOverlaps}
\alias{calculateOverlaps}
\title{Calculates percent overlap of IUCN and NHM polygons}
\usage{
calculateOverlaps(x, y)
}
\arguments{
\item{x}{an object of class sf, sfc or sfg containing Natural History Museum
records}

\item{y}{object of class sf, sfc or sfg, containing IUCN range polygon data}
}
\value{
object \code{x}, the NHM data set, with an added percent overlap
column, containing calculated percentage overlap between NHM and
IUCN range polygons
}
\description{
\code{calculateOverlaps} calculates the percentage overlap between the
Natural History Museum (NHM) and IUCN range polygons.
}
\details{
This is a two input function for calculating the percentage overlap between
NHM and IUCN datasets, where both are polygons.
}
\examples{
NHM_Data <- hullOverlaps(NHM_Data, IUCN_Data)
}
