% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{cent}
\alias{cent}
\title{Compute centroid in one dimension (X or Y)}
\usage{
cent(x, dim, crs = NA)
}
\arguments{
\item{x}{sf object}

\item{dim}{dimension, 'X' or 'Y'}

\item{crs}{coordinate reference system (CRS) to use for geometric operations.
Necessary when using a geographic CRS; can be an epsg code}
}
\value{
numeric of centroid in one dimension
}
\description{
Compute centroid in one dimension (X or Y)
}
