% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroidEdgeDistance.R
\name{centroidEdgeDistance}
\alias{centroidEdgeDistance}
\title{Calculates the distance between the centroid of a one convex hull and up to
two other hulls}
\usage{
centroidEdgeDistance(x, y)
}
\arguments{
\item{x}{an object of class sf, sfc or sfg containing data from Natural
History Museum archives. Contains a \code{binomial} column and geometry.}

\item{y}{n object of class sf, sfc or sfg containing data from IUCN records.
Contains a \code{binomial} column and geometry.}
}
\value{
returns \code{x} now with additional \code{distance} columns, which
contain the computed distance between the centroid given in \code{x} and the
nearest edge of the centroid(s) given in \code{y}
}
\description{
\code{centroidEdgeDistance} calculates the distance from the centroid of one
convex hull to the edge of up to two other hulls for a given species. This
was developed to produce data for further analysis of species range changes
from historical records to current estimates.
}
\examples{
x <- centroidEdgeDistance(x, y)
}
