% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipHullsToLand.R
\name{clipHullsToLand}
\alias{clipHullsToLand}
\title{Clips convex hulls to landmasses}
\usage{
clipHullsToLand(x)
}
\arguments{
\item{x}{object of class sf, sfc or sfg}
}
\value{
Returns object \code{x} with clipped convex hulls computed and stored
in the \code{convex_hull} column.
}
\description{
\code{clipHullsToLand} returns the clipped convex hulls by species for a
single object of class sf (a data frame).
}
\details{
These hulls are clipped to landmasses only.
To generate hulls and clip them to landmasses with a single function see
\link[sfe]{makeLandClippedHulls}. \code{x} must have
both a \code{geometry} and a \code{binomial} column present to work.

For more felxible convex hull generation see \link[sf]{st_convex_hull}

Landmasses used to clip to are generated from the \code{rnaturalearth}
package using the \code{ne_countries} function. please see its documentation
here \link[rnaturalearth]{ne_countries} for a full breakdown.
}
\examples{
data.frame <- clipHullsToLand(data.frame)
}
