% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colnames.R
\name{correctColName}
\alias{correctColName}
\title{Renames the desired column to 'binomial'}
\usage{
correctColName(x)
}
\arguments{
\item{x}{object of class sf, sfc or sfg}
}
\value{
An sf data frame with a column renamed as 'binomial'
}
\description{
\code{correctColName} returns the input data frame with whichever column
was sepcified in the console renamed to \code{binomial}.
}
\details{
Many other functions require a binomial column to work, as they subset by
species.
}
\examples{
data.frame <- correctColName(data.frame)
}
