% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countOverlaps.R, R/countsForAnalysis.R
\name{countNumbOverlaps}
\alias{countNumbOverlaps}
\title{Counts the number of binomial overlaps for each species in a data frame}
\usage{
countNumbOverlaps(x)

countNumbOverlaps(x)
}
\arguments{
\item{x}{an object of class sf, sfc or sfg containing collection records,
including binomial names and binomial overlaps, which can be calculated using
\code{sfe::binomialOverlap()}.}

\item{x}{an object of class sf, sfc or sfg containing collection records,
including binomial names and binomial overlaps, which can be calculated using
\code{sfe::binomialOverlap()}.}
}
\value{
\code{x} with a new column of binomial overlap counts added.

\code{x} with a new column of binomial overlap counts added.
}
\description{
\code{countNumbOverlaps} counts the number of binomial overlaps for each
species in a data frame. Creates a new column \code{numberOfOverlaps} with
this count data for each species.

\code{countNumbOverlaps} counts the number of binomial overlaps for each
species in a data frame. Creates a new column \code{numberOfOverlaps} with
this count data for each species.
}
\examples{
x <- countNumbOverlaps(x)

x <- countNumbOverlaps(x)

}
