% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsForAnalysis.R
\name{countNumbSpecimens}
\alias{countNumbSpecimens}
\title{Counts the number of specimens for each species in a data frame}
\usage{
countNumbSpecimens(x)
}
\arguments{
\item{x}{an object of class sf, sfc or sfg containing collection records,
including binomial names.}
}
\value{
\code{x} with a new column of specimen counts added.
}
\description{
\code{countNumbSpecimens} counts the number of unique specimens for each
species in a data frame. Creates a new column \code{numberOfSpecimens} with
this count data for each species.
}
\examples{
x <- countNumbSpecimens(x)

}
