% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixTypeNames.R
\name{fixTypeNames}
\alias{fixTypeNames}
\title{Cleans TypeStatus column so that variants of TypeStatus are unified.}
\usage{
fixTypeNames(x)
}
\arguments{
\item{x}{an object of class sf, sfc or sfg containing a \code{TypeStatus}
column.}
}
\value{
an object of class sf, sfc or sfg containing a unified TypeStatus
column
}
\description{
Uses regex to replace type denotations with the prefix only, unless the
denotation is type (or variant), in which case it replaces the variants with
\code{type}
}
\details{
\code{fixTypeNames} corrects variants of type description (e.g. Type, Types,
type, etc.) and turns them into a single description.
}
\examples{
data.frame <- fixTypeNames(data.frame)
}
