% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{gen_features}
\alias{gen_features}
\title{Create features from one user response}
\usage{
gen_features(file, true.loc, labels, crs = 4326)
}
\arguments{
\item{file}{path to file as character string}

\item{true.loc}{sf object of true locations}

\item{labels}{character vector of short labels to use for locations; needed
for generating succinct column names for features}

\item{crs}{coordinate reference system to use for distance calculations}
}
\description{
Create features from one user response
}
