% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{get_bearing}
\alias{get_bearing}
\title{Get bearing (angle) between -180 and 180 from one set of points (i.e., responses) to another (i.e. actual locations)}
\usage{
get_bearing(set1, set2)
}
\arguments{
\item{set1}{sf object of locations; must have 'lng' and 'lat' columns}

\item{set2}{sf object of locations; must have 'lng' and 'lat' columns}
}
\description{
Get bearing (angle) between -180 and 180 from one set of points (i.e., responses) to another (i.e. actual locations)
}
