% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{get_dist_geo}
\alias{get_dist_geo}
\title{function for getting geodesic distance}
\usage{
get_dist_geo(p1, p2)
}
\arguments{
\item{p1}{set of points of class sf}

\item{p2}{set of points of class sf}
}
\description{
function for getting geodesic distance
}
