% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inReadOld.R
\name{inReadOld}
\alias{inReadOld}
\title{Reads in old format geospatial data from
User terminal input
Terminal prompts ask for path to data and layer type}
\usage{
inReadOld(path = NULL, name = NULL)
}
\arguments{
\item{path}{Path to data entered as a string, default is \code{NULL}}

\item{name}{layer for readOGR to work with, entered as a string,  default is
\code{NULL}

Parameters are enterd into the console where prompted. Both a
path to an IUCN data set and then the layer to read need to be entered}
}
\value{
'IUCNData', now in sf format
}
\description{
Reads in old format geospatial data from
User terminal input
Terminal prompts ask for path to data and layer type
}
\examples{
inReadOld()
"enter path:"
../Data/IUCN_Data
"enter layer:"
PANGOLINS
}
