% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{load_one_response}
\alias{load_one_response}
\title{function for testing and getting colnames needed for labels}
\usage{
load_one_response(file, labels, crs = 4326)
}
\arguments{
\item{file}{path to file as character string}
}
\description{
function for testing and getting colnames needed for labels
}
