% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{load_polygons}
\alias{load_polygons}
\title{Load polygon data as single sf dataset}
\usage{
load_polygons(dsn)
}
\arguments{
\item{dsn}{data source directory}
}
\description{
Load polygon data as single sf dataset
}
