% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeConvexHulls.R
\name{makeConvexHulls}
\alias{makeConvexHulls}
\title{Creates Convex Hulls}
\usage{
makeConvexHulls(x)
}
\arguments{
\item{x}{object of class sf, sfc or sfg}
}
\value{
Returns object \code{x} with convex hulls computed and stored in the
\code{geometry} column.
}
\description{
\code{makeConvexHulls} returns the computed convex hulls by species for a
single object of class sf (a data frame).
}
\details{
These hulls are not clipped,
for hulls clipped to landmasses, see \link[sfe]{clipHullsToLand},
or alternatively see \link[sfe]{makeLandClippedHulls}. \code{x} must have
both a \code{geometry} and a \code{binomial} column present to work.

For more felxible convex hull generation see \link[sf]{st_convex_hull}
}
\examples{
data.frame <- makeConvexHull(data.frame)
}
