% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLandClippedHulls.R
\name{makeLandClippedHulls}
\alias{makeLandClippedHulls}
\title{Creates convex hulls which are clipped to landmasses}
\usage{
makeLandClippedHulls(x)
}
\arguments{
\item{x}{object of class sf, sfc or sfg}
}
\value{
An object (\code{x}) of class sf, sfc or sfg with clipped convex
hulls computed and stored in the \code{geometry} column.
}
\description{
\code{makeLandClippedHulls} returns the clipped convex hulls by species for a
single object of class sf (a data frame).
}
\details{
CHECK THIS I MIGHT HAVE FIXED THAT

A warning message will appear for \code{st_intersection}, which is correct,
however so long as the crs of your \code{geometry} column is
longitude/latitude (4326) it is of no consequence.

These hulls are clipped to landmasses only. \code{x} must have
both a \code{geometry} and a \code{binomial} column present to work.

For more felxible convex hull generation see \link[sf]{st_convex_hull}

Landmasses used to clip to are generated from the \code{rnaturalearth}
package using the \code{ne_countries} function. please see its documentation
here \link[rnaturalearth]{ne_countries} for a full breakdown.
}
\examples{
data.frame <- makeLandClippedHulls(data.frame)
}
