% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchBinomial.R
\name{matchBinomial}
\alias{matchBinomial}
\title{Finds matching binomials (records) between NHM and IUCN data frames}
\usage{
matchBinomial(x, y)
}
\arguments{
\item{x}{An object of class sf containing NHM data with binomial column}

\item{y}{An object of class sf containing IUCN data with binomial column}
}
\value{
\code{x}, an object of class sf of NHM data, filtered to contain
records binomial which are also present in the IUCN data set.

\code{IUCNData} an object of class sf containing IUCN data, filtered
to contain binomial records present in NHM data
}
\description{
\code{matchBinomial} reduces both data frames to contain only records which
both contain. Aims to aid with the speed of subsequent calculations.
}
\examples{
matchBinomial()
}
