% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiedBinomialOverlap.R
\name{modifiedBinomialOverlap}
\alias{modifiedBinomialOverlap}
\title{Turns percentage overlap data into binomial data at a 50\% threshold}
\usage{
modifiedBinomialOverlap(x)
}
\arguments{
\item{x}{an object of class sf, sfc or sfg, containing calculated percentage
overlaps of range polygons between Natrual History Museum data and IUCN data.}
}
\value{
\code{x} with the \code{Percent_overlaps} column "flattened" to
binomial data, and stored in the \code{binomial_overlap} column
}
\description{
\code{modifiedBinomialOverlap} turns previously calculated percentage overlap data
into binomial (yes/no or 1/0) overlap data on a per species basis. This is in
preparation for analysis using binomial GLMM's.
}
\details{
You will get a warning messages saying only the first value will be used to
check if an \code{if} condition is true, these arise due to R interpreting
the column of \code{Percent_overlap} as a list, so can be safely ignored.
}
\examples{
x <- modifiedBinomialOverlap(x)
}
