% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{num_points}
\alias{num_points}
\title{Count number of points comprising polygon}
\usage{
num_points(x)
}
\arguments{
\item{x}{sf object}
}
\value{
int
}
\description{
Count number of points comprising polygon
}
