% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMaps.R
\name{plotMaps}
\alias{plotMaps}
\title{Plots maps of each species in a binomial list}
\usage{
plotMaps(x, y, path = NULL)
}
\arguments{
\item{x}{An object of class sf, sfc or sfg containing Natural History Museum
records}

\item{y}{An object of class sf, sfc or sfg containing IUCN records}

\item{path}{path to destination folder entered as a string. This must have
been created before using this function otherwise you will get an error.}
}
\value{
plots in the dir specified by user input
}
\description{
\code{plotMaps} plots maps showing both IUCN and Natural History Museum data
to a directory specified by the user. The directroy must have been created
prior to using this function, otherwise a "no such directory exists" error is
thrown.
}
\examples{
plotMaps(x, y)
'Enter path: path/to/dir'
}
