% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{point_features}
\alias{point_features}
\title{Create features from a directory of points}
\usage{
point_features(dsn, true.loc, labels, crs = 4326)
}
\arguments{
\item{dsn}{character string of data directory from where .csv's are stored}

\item{true.loc}{sf object preferably generated from the function
true_locations, used for calculating distance from user response to the
true locations}

\item{labels}{character vector of short labels to use for locations; needed
for generating succinct column names for features}

\item{crs}{coordinate reference system to use for distance calculations}
}
\description{
Create features from a directory of points
}
