% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{poly_features}
\alias{poly_features}
\title{Feature engineering for polygons}
\usage{
poly_features(x, crs)
}
\arguments{
\item{x}{sf object}

\item{crs}{crs code (numeric) to use for geometric operations}
}
\value{
sf object with original columns and newly created columns
}
\description{
Feature engineering for polygons
}
