% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepNHMData.R
\name{prepNHMData}
\alias{prepNHMData}
\title{Reads a .csv file and converts it to a filtered sf object (a data frame)}
\usage{
prepNHMData(x, indexToChange = NULL)
}
\arguments{
\item{x}{a csv file containing NHM records}

\item{indexToChange}{if the index you need to change is already known please
enter it here to save time. Including this mostly to make the vignette run
properly as it doesn't like terminal input. Default of \code{NULL} as in
most cases I would assume the index is unknown and this function works fine
outside of the vignette.}
}
\value{
An object of class sf (a data frame) with a column renamed to
\code{binomial}, with binomial names underscorred, filtered to exclude NA
values.
}
\description{
\code{prepNHMData} Reads in old format geospatial data taking user input.
Terminal prompts ask for path to data and layer type, once entered
prepIUCNData is read in and reformated to 'sf' format. Spaces in the
\code{binomial} column are replaced with \code{_}
}
\examples{
data.frame <- prepNHMData()
'please enter path'
path/to/your/data
'please enter column index to rename'
YourIndexHere

}
