% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolveIUCNGeom.R
\name{resolveIUCNGeom}
\alias{resolveIUCNGeom}
\title{Unifies IUCN range records to a single entry for each species}
\usage{
resolveIUCNGeom(x)
}
\arguments{
\item{x}{an object of class sf, sfc or sfg containing IUCN records, where
each species has multiple entries.}
}
\value{
object \code{output}, the IUCN data set, now with a single entry for
each species, where the geometry has been unified but internal boundaries have
not been resolved, to retain information about range fragmentation.
}
\description{
\code{resolveIUCNGeom} turns IUCN data which has multiple entries for each
species into a single record for each species, so that percentage overlaps
can be calculated. Internal boundaries are not resolved to preserve
information about range fragmentation.
}
\examples{
TestOut <- resolveIUCNGeom(IUCNForPlot)
}
