% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SFE-package.R
\docType{package}
\name{SFE-package}
\alias{SFE}
\alias{SFE-package}
\title{SFE: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Shuangbin Xu \email{xshuangbin@163.com} (\href{https://orcid.org/0000-0003-3513-5362}{ORCID})

}
\keyword{internal}
