% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{to_wide}
\alias{to_wide}
\title{function for making data wide, getting x/y as columns specific to each location}
\usage{
to_wide(df, labels)
}
\description{
function for making data wide, getting x/y as columns specific to each location
}
