% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{true_locations}
\alias{true_locations}
\title{Create sf object of true locations based on labels, x, and y}
\usage{
true_locations(labels, lng, lat, crs = 4326)
}
\arguments{
\item{labels}{character vector of short labels to use for locations; needed
for generating succinct column names for features}

\item{crs}{coordinate reference system to use for distance calculations}

\item{x}{longitude coordinates of true locations (numeric vector)}

\item{y}{latitude coordinates of true locations (numeric vector)}
}
\description{
Create sf object of true locations based on labels, x, and y
}
