% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateVersion.R
\name{updateVersion}
\alias{updateVersion}
\title{Updates the version number of uploaded data via console input}
\usage{
updateVersion()
}
\value{
Updated DESCRIPTION file for easy versionning
}
\description{
\code{updateVersion} updates the version number (GitHub release) via user
input. The console will ask that the user specify a new version number, which
will be refelcted in the DESCRIPTION file. This removes the need to hunt for
the DESCRIPTION FILE and manually open/edit/save/close it before committing
a newly edited data set.
}
\examples{
updateVersion()
'please enter new version number:'
Your.New.Version.Number
}
