% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepIUCNData.R
\name{prepIUCNData}
\alias{prepIUCNData}
\title{Prepares IUCN data for analysis using sf and sfe functions}
\usage{
prepIUCNData(path = NULL)
}
\arguments{
\item{path}{is the path to IUCN data}
}
\value{
sf format IUCN data, with binomial underscored
}
\description{
\code{prepIUCNData} Reads in old format geospatial data taking user input.
Terminal prompts ask for path to data and layer type, once entered
prepIUCNData is read in and reformated to 'sf' format. Spaces in the
\code{binomial} column are replaced with \code{_}
}
\examples{
data.frame <- prepIUCNData()
'please enter path'
path/to/your/data
'please enter layer'
AMPHIBIANS

}
